fix: user axios 리팩터링(이전커밋) 오류 수정
