+++ src/pages/CallBack.tsx
@@ -25,7 +25,6 @@ const CallBack = () => {
           res.data.member_imgUrl,
           res.data.oauthId,
         );
-        // console.log(res);
       })
       .catch((err) => {
         console.log(err);
